Module ptree::print_config

Structures to control the output formatting

Output formatting is configured through the PrintConfig structure.

Structs

IndentChars

Set of characters use to draw indentation lines (branches)

PrintConfig

Structure controlling the print output formatting

StaticIndentChars

Set of characters use to draw indentation lines (branches)

Enums

OutputKind

Enumeration of output kinds

StyleWhen

Configuration option controlling when output styling is used

Constants

ASCII_CHARS_PLUS

ASCII indentation characters, using a plus (+) for turning right

ASCII_CHARS_TICK

ASCII indentation characters, using a tick (```) for turning right

UTF_CHARS

UTF-8 indentation characters, using regular box-drawing characters

UTF_CHARS_BOLD

UTF-8 indentation characters, using heavy box-drawing characters

UTF_CHARS_DASHED

UTF-8 indentation characters, using dashed box-drawing characters

UTF_CHARS_DOUBLE

UTF-8 indentation characters, using double box-drawing characters